took out the General and Specific Reminder functions from the index file and placed it into it's own file because the iisnode and the forever node package running the index file, in prod, was causing duplicate reminders to be sent to users.
